Fix: normalize cluster threshold to [0,1] RF distance, matching compare_trees_rf()
Agent-Logs-Url: https://github.com/ChrisCreevey/clann/sessions/91a5b801-dc4f-491e-8b03-9933e832b4df

+++ treecluster.h
@@ -31,7 +31,10 @@
  *
  *   lm         : fully-populated LandscapeMap (must be non-NULL)
  *   out_file   : output TSV filename
- *   threshold  : max RF distance to join a cluster (>= 0)
+ *   threshold  : max normalized RF distance to join a cluster (0.0–1.0).
+ *                Mirrors compare_trees_rf(): normalized by 2*(n-3) where
+ *                n is the number of taxa in the tree.
+ *                e.g. 0.1 = within 10% of the maximum possible RF distance.
  *   orderby    : 0 = sort by score ascending (best first),
  *                1 = sort by visit_count descending (most-visited first)
  *
@@ -45,7 +48,7 @@
  */
 void lm_cluster(LandscapeMap *lm,
                 const char   *out_file,
-                int           threshold,
+                float         threshold,
                 int           orderby);
 
 #endif /* CLANN_TREECLUSTER_H */
